move matrix stack to renderer
