+++ cocos/renderer/CCRenderer.cpp
@@ -171,6 +171,163 @@ void RenderQueue::restoreRenderState()
     CHECK_GL_ERROR_DEBUG();
 }
 
+
+MatrixStack::MatrixStack()
+{
+    initMatrixStack();
+}
+
+MatrixStack::~MatrixStack()
+{
+    
+}
+
+void MatrixStack::initMatrixStack()
+{
+    while (!_modelViewMatrixStack.empty())
+    {
+        _modelViewMatrixStack.pop();
+    }
+    
+    while (!_projectionMatrixStack.empty())
+    {
+        _projectionMatrixStack.pop();
+    }
+    
+    while (!_textureMatrixStack.empty())
+    {
+        _textureMatrixStack.pop();
+    }
+    
+    _modelViewMatrixStack.push(Mat4::IDENTITY);
+    _projectionMatrixStack.push(Mat4::IDENTITY);
+    _textureMatrixStack.push(Mat4::IDENTITY);
+}
+
+void MatrixStack::resetMatrixStack()
+{
+    initMatrixStack();
+}
+
+void MatrixStack::loadIdentityMatrix(MATRIX_STACK_TYPE type)
+{
+    if(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW == type)
+    {
+        _modelViewMatrixStack.top() = Mat4::IDENTITY;
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION == type)
+    {
+        _projectionMatrixStack.top() = Mat4::IDENTITY;
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_TEXTURE == type)
+    {
+        _textureMatrixStack.top() = Mat4::IDENTITY;
+    }
+    else
+    {
+        CCASSERT(false, "unknow matrix stack type");
+    }
+}
+
+void MatrixStack::loadMatrix(MATRIX_STACK_TYPE type, const Mat4& mat)
+{
+    if(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW == type)
+    {
+        _modelViewMatrixStack.top() = mat;
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION == type)
+    {
+        _projectionMatrixStack.top() = mat;
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_TEXTURE == type)
+    {
+        _textureMatrixStack.top() = mat;
+    }
+    else
+    {
+        CCASSERT(false, "unknow matrix stack type");
+    }
+}
+
+void MatrixStack::multiplyMatrix(MATRIX_STACK_TYPE type, const Mat4& mat)
+{
+    if(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW == type)
+    {
+        _modelViewMatrixStack.top() *= mat;
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION == type)
+    {
+        _projectionMatrixStack.top() *= mat;
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_TEXTURE == type)
+    {
+        _textureMatrixStack.top() *= mat;
+    }
+    else
+    {
+        CCASSERT(false, "unknow matrix stack type");
+    }
+}
+
+void MatrixStack::pushMatrix(MATRIX_STACK_TYPE type)
+{
+    if(type == MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW)
+    {
+        _modelViewMatrixStack.push(_modelViewMatrixStack.top());
+    }
+    else if(type == MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION)
+    {
+        _projectionMatrixStack.push(_projectionMatrixStack.top());
+    }
+    else if(type == MATRIX_STACK_TYPE::MATRIX_STACK_TEXTURE)
+    {
+        _textureMatrixStack.push(_textureMatrixStack.top());
+    }
+    else
+    {
+        CCASSERT(false, "unknow matrix stack type");
+    }
+}
+
+void MatrixStack::popMatrix(MATRIX_STACK_TYPE type)
+{
+    if(MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW == type)
+    {
+        _modelViewMatrixStack.pop();
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION == type)
+    {
+        _projectionMatrixStack.pop();
+    }
+    else if(MATRIX_STACK_TYPE::MATRIX_STACK_TEXTURE == type)
+    {
+        _textureMatrixStack.pop();
+    }
+    else
+    {
+        CCASSERT(false, "unknow matrix stack type");
+    }
+}
+
+const Mat4& MatrixStack::getMatrix(MATRIX_STACK_TYPE type)
+{
+    if(type == MATRIX_STACK_TYPE::MATRIX_STACK_MODELVIEW)
+    {
+        return _modelViewMatrixStack.top();
+    }
+    else if(type == MATRIX_STACK_TYPE::MATRIX_STACK_PROJECTION)
+    {
+        return _projectionMatrixStack.top();
+    }
+    else if(type == MATRIX_STACK_TYPE::MATRIX_STACK_TEXTURE)
+    {
+        return _textureMatrixStack.top();
+    }
+    
+    CCASSERT(false, "unknow matrix stack type, will return modelview matrix instead");
+    return  _modelViewMatrixStack.top();
+}
+
 //
 //
 //
